GeoLearning/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.2.1
autopep8==2.0.0
Babel==2.12.1
backcall==0.2.0
bokeh==3.1.0
calmap==0.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
Django==3.2.3
django-admin-interface==0.20.0
django-ckeditor==6.7.0
django-colorfield==0.7.2
django-embed-video==1.4.7
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.1.0
django-nested-admin==4.0.2
django-scheduler==0.10.1
django-widget-tweaks==1.5.0
entrypoints==0.4
executing==1.2.0
fonttools==4.41.0
greenlet==2.0.2
icalendar==5.0.7
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
july==0.1.3
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
Mako==1.3.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==23.0
pandas==1.5.2
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
pygame==2.5.2
Pygments==2.13.0
pygubu==0.31
pygubu-designer==0.36
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
pyqtgraph==0.13.3
PySide2==5.15.2.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2022.4
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==24.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.28.1
screeninfo==0.8.1
seaborn==0.12.2
shiboken2==5.15.2.1
shiboken6==6.5.1.1
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.3
stack-data==0.6.2
tkcalendar==1.6.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.5
xyzservices==2023.2.0
